+++ Changelog.md
@@ -2,6 +2,7 @@ v0.2
 - [ApWorld] Fixed minor logic bugs
   - `Napalm Charge` not requiring  `Spinning Widget`
   - `N-Step Branch Prediction` requiring `Ascended Widget` instead of `Unshackled Widget`
+- [ApWorld] Added a yaml setting for starting hints for producers upto a specific tier
 - [Client] Added Custom Assets support
 - [Client] Implemented scouting locations
   - status messages (top right) now displays sent/recieved items